Force gcc >= 3.4 for certain compiler attributes.
It was reported to me that Tuvok compilation failed with gcc 3.3.
I think this was because some of the attribute syntax I was using
was introduced in 3.4, so force that version before using those
attributes.

In general this isn't a great approach, because 3.3 and earlier
users still get the wrong behavior .. but we're (currently) only
using this for error suppression.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@645 c36c8488-0289-0348-9b64-b301f74bd9a7