Try harder to get the amount of system memory.
Attempt a getrlimit(2) to read the data size, since sysinfo seems
to fail in strange cases.  Unfortunately getrlimit has a habit of
telling us we have infinite memory, so we still fall back on
sysinfo in most cases.

More importantly, when either call fails, we now print an error
message with the reason.  We don't have access to the debug logs
here, so it will always go to the users terminal, but that's still
better than nothing.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@669 c36c8488-0289-0348-9b64-b301f74bd9a7