Pass world and view matrices into the SB geometry generator.
This commit gives both the world coordinate transformation and the
view transformation to the geometry generator.  Doing so allows it
to distinguish between the two matrices, and only apply the view
transformation to the embedded clip plane.

We've also overridden (Enable|Disable)ClipPlane in the OpenGL
Slice-based VRer, so that it has knowledge of when this is
changed.  This gives the renderer enough knowledge to update the
geometry whenever the clip plane is changed, instead of every
frame.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@791 c36c8488-0289-0348-9b64-b301f74bd9a7