Add TIFF header files to 3rdParty.
Qt has tiff builtin, but the installed Qt won't have the plugin
header files.  Thus, this satisfies the compile-time dependency,
but we rely on Qt to satisfy the link-time dependency.

These headers come from libtiff-3.8.2, which appears to be the
same version used in Qt-4.4.3 (and also the current release).

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@832 c36c8488-0289-0348-9b64-b301f74bd9a7