Get rid of a bogus comment, and initialize the samples variable.
Seems as if libtiff will conveniently leave the variable
unmodified if the tag isn't found in the file, or maybe if reading
the tag gives some other sort of error.  The example I have does
have a valid tag, but would give bogus results reading it, only on
the Mac for some reason.

Anyway, this code correctly reads in both 8 and 16 bit datasets.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@852 c36c8488-0289-0348-9b64-b301f74bd9a7