Add a virtual method for scaling the transfer function.
Client apps that give us a dataset plus transfer function tend to
do so in a `raw' fashion -- any transformations on that data (such
as making the TF range match the scalar range) are already done.

This allows derived renderers to override the implicit scaling
that Tuvok does to the TFs, while still taking advantage of code
to initialize shader vars.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@958 c36c8488-0289-0348-9b64-b301f74bd9a7