Beginnings of a more general design for data access.
Existing API is difficult to work with when there is no UVF file
backing the data.  This starts a new hierarchy where a UVF dataset
is a child of a more general dataset.

Eventually, by deriving from the superclass, we can provide an
easier method to insert data from external sources into Tuvok's
rendering subsystem.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@996 c36c8488-0289-0348-9b64-b301f74bd9a7