Add tiff to vcproj.
Adds tiff implementation files to vcproj, to make sure we're accessing
the library uniformly on all platforms.

Adds in a couple hacks to the tiff implementation to fix compile
troubles.  `inline' isn't standard in C, and it's just an efficiency
issue anyway so we just remove it.  Secondly, we lack anything like a
`configure' step, so just explicitly say win32 doesn't support mmap.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1138 c36c8488-0289-0348-9b64-b301f74bd9a7