Removed unused bricks before falling back to harsher measures.
Running out of memory like this is pretty harsh.  Unlike hitting
*our* memory limit, the out of memory condition here implies the
system really is strained for memory.  Most memory allocators
start taking exponential time to find a free block when that
occurs.

For that reason, we really want to get rid of a *lot* of bricks.
Deleting bricks which are `next' to each other in the allocator's
lists relieves pressure on the OS' memory manager, which is having
trouble finding contiguous blocks large enough for our bricks.

Despite this new strategy, there's still a buglet that we'll
delete bricks which are still in use if it's absolutely necessary.
The alternative is to crash, or maybe abort the render entirely.
I kept this behavior so that we could potentially implement some
sort of more advanced scheme later, whereby the memory manager
informs a higher level, `I know you were using this brick, but I
had to delete it; reallocate it if you still need it.'.

With this commit, I can render the 2.9GB `magnitude' dataset on a
32-bit Mac build which has 2GB of RAM.  In rare circumstances, a
brick ends up missing at the highest LoD, seemingly due to the
deletion of a brick which is in use.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1184 c36c8488-0289-0348-9b64-b301f74bd9a7