GPU-MM functionality for external DS's.
Create 'AddDataset' and make 'FreeAssociatedTextures' public.
This allows the renderer to add datasets, plus cleanup 3D textures
when new data are uploaded.  External datasets are a special case
in this respect: data often change at runtime, and when that
occurs any old textures are useless -- we're guaranteed they'll
never be used again.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1267 c36c8488-0289-0348-9b64-b301f74bd9a7