Store/manipulate isovalues `natively'.
That is, AbstrRenderer's isovalue is in the range of the data.
The UI needs integers, so it is responsible for that conversion;
the GL renderer needs data in the range [0,1], so it does that
conversion (that said, a convenience method,
GetNormalizedIsovalue, was added for the latter conversion).

The isovalue defaults to half the data range (as before).

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1326 c36c8488-0289-0348-9b64-b301f74bd9a7