Improve build, particular on OS X.
Don't force any particular architecture; rely on the Qt install in
use to appopriately set the architecture[s].  Thus, if you want a
64bit Tuvok, run `qmake' from a 64bit Qt install.  The one issue is
that, on OS X, you need to change the QT_MAC_USE_COCOA define in
the project file from 0 to 1.

Also add tiff to our list of plugins; users seem to be having
issues loading a directory full of TIFFs.  Not sure why it hasn't
been there all along...


git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1340 c36c8488-0289-0348-9b64-b301f74bd9a7