Fix crash loading bad GLSL shaders on some systems.
If we can't load the shaders, we attempt to query the driver as to
what went wrong.  The query call, if it fails, is defined not to
modify the `length' parameter we are asking about.  The automatic
variable we were using is uninitialized but normally 0, but can get
any value -- including absurd values that cause a future 'new' to
fail.

Now we explicitly initialize the variable, so if the GL call fails
we'll recognize that something is wrong.


git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1342 c36c8488-0289-0348-9b64-b301f74bd9a7