Fix a minor issue which made Tuvok choose the wrong LOD.
The 3 frame allowance to "catch up" is great if you're actively
rotating the volume and things slow down a bit.  However, if
you've *just* changed the rendering parameters and the very first
frame can't render at the given FPS, we want to jump down to the
coarsest LOD.

Turns out that we'll think the finer LOD takes 0.0f seconds in
that case.  By keying into that we can leave the "catch up" code
in place for now.

This also breaks a lot of long lines in the LOD choice code.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1349 c36c8488-0289-0348-9b64-b301f74bd9a7