First stage of removing explicit 2x2 code from Tuvok into client.  In the end Tuvok will just render RenderRegions and will not care how many there are, what they are (there could be 5 sagittal views on the screen), or where they are (ERenderArea will go away).
This first pass introduces some of the infrastructure and in the
process makes Tuvok a bit cleaner.  It does not require any changes to
the client (which makes the code still a bit ugly).  Future commits
will require changes to the client but make Tuvok way simpler and more
powerful.


git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1366 c36c8488-0289-0348-9b64-b301f74bd9a7