Partial audit to use %u and unsigned.
Many of our internal values are size_t's, or UINT?? types.  There
is no portable way to print size_t's.  We've been casting them to
ints, but we can get an extra bit of precision if we use unsigned
values, and switch from "%d" or "%i" to "%u" in our format
specifiers.

This does a partial audit aimed at using unsigned values when
possible.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1404 c36c8488-0289-0348-9b64-b301f74bd9a7