Detect errors when loading 2D TFqns.
Unfortunately we just debugout an error message and bail, loading
up an empty TF.  Thus a user won't get a real notification that
their TFqn is corrupt.

The error really needs to make it up to the UI.  In some sense,
that means an exception, because that propagates easiest.
However, failure to load an "automatic" TF shouldn't prevent
loading a dataset, so we would at least need a boolean to check
for failure asynchronously, for that case.

We'll implement it that way eventually.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1487 c36c8488-0289-0348-9b64-b301f74bd9a7