Fix a Mac-only bug while detaching shaders.
We query GL for the list of shaders attached to a program object,
so that we can detach those shaders.  The Mac GL implementation is
deciding that shader 0 is attached to our program, which can't
possibly make sense because shader 0 is not a real shader.

So, before detaching shaders, we'll check if they are actually a
shader.  This means we ask GL for a list of shaders, and then we
need to ask GL if each entry in the list is a shader.


git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1538 c36c8488-0289-0348-9b64-b301f74bd9a7