Rendering support for RGBA data.
Co-opts the 1D TFqn mode to render the data as-is.  2D TFqns are
still 'broken' (renders the 'r' component, no histograms), but
they've always been like that anyway.

This works by compiling one of vr-std or vr-no-tfqn depending on
whether or not we find the data set is RGBA or not.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1717 c36c8488-0289-0348-9b64-b301f74bd9a7