Guard against error states.
If we get into an error state, we can get INT_MAX for log_length.
This either comes about because a driver is braindead or log_length
happens to take that value because it was uninitialized; however it
comes about, a gigantic value causes the vector initialization to
take forever.


git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1737 c36c8488-0289-0348-9b64-b301f74bd9a7