Fix isosurface rendering mode, pre-resize.
The issue was that we wouldn't StartFrame() after scheduling a
complete redraw.  This checks if all our regions are blank, as
happens when we schedule a full redraw (which, in turn, we do
whenever switching rendering modes, of course); if they're all
blank, we need to start a new frame.

I chose to look at 'isBlank', as opposed to 'redrawMask' or
'isTargetBlank'... mostly because I no longer remember what all
these variables mean.  We need to simplify some of this state.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1827 c36c8488-0289-0348-9b64-b301f74bd9a7