Don't always enable debug stl.
It will break at link time if libTuvok.a was not built with debug
stl as well.  By simply adding the flag during qmake-time, it
should cause /all/ projects to use the flag, and thus it ensures
we're consistent... at the risk of forgetting to enable it in some
cases.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1891 c36c8488-0289-0348-9b64-b301f74bd9a7