Forget PRL files.
They have created nothing but hassle.  I hate that when someone
links "Tuvok" they need to link "tuvokexpr" as well, and that'll
get worse when we actually get a scio library... but it's just not
worth it.  They are just plain broken; always adding paths to our
link lines that linkers complain about.  On windows, it creates a
fatal error.

Adding the additional link library and directories in explicitly
sucks, but it actually isn't worse than dealing with the braindead
implementation of Qt's PRL files.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1914 c36c8488-0289-0348-9b64-b301f74bd9a7