Remove hack for intel uniform IDs.
While the uniforms we get back /are/ actually pretty crazy, they
are not wrong.  The spec allows an implementation to use whatever
the hell it wants for these.  Intel must be using an internal hash
table or something instead of just a list, unlike other drivers,
since the values we get back are so wild.

Anyway, hacking the uniform ID actually breaks things.  Let it be.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@1933 c36c8488-0289-0348-9b64-b301f74bd9a7