Add function to set an isovalue relatively.
We want this when we have locked views which we change an
isovalue for.  If the two data sets have the same bit width,
whatever, no problem. However, if the data sets have different bit
widths, setting an isovalue of "4823" for example might not make
sense (e.g 8bit data).

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2149 c36c8488-0289-0348-9b64-b301f74bd9a7