Consistently use API for testing if we have color data.
Rename "RGBAData" to "ColorData", and use it throughout, instead
of manually checking component count. In the long term, we'll want
a more complicated check that queries the data set anyway. That is,
the data set might want to store 4-component data but say that it
isn't color data for whatever reason.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2165 c36c8488-0289-0348-9b64-b301f74bd9a7