Fixed Lua stack error.
The ClassConstructor::exec function was returning an incorrect type
because LuaStackRAII popped its return values off the stack.

May need to re-implement LuaStackRAII so that if values
are being returned on the stack, then valid return values are not popped
off the stack. Instead we pop the values starting at stack
position Top-(n-1).

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2237 c36c8488-0289-0348-9b64-b301f74bd9a7