Add some hacky stuff for playing with render state.
Allows querying and setting render state.  The only real
query right now is looking up frame times.  But we can now set the
missing brick strategy for the Tree Raycaster *at*runtime* via
Lua.

Some of this is pretty gross.  I've taken to the convention of
prefixing things with "PH_" if they are (P)aper (H)acks which we
should outright delete or at the very least rethink after we
finish up the benchmarking.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2402 c36c8488-0289-0348-9b64-b301f74bd9a7