Added CaptureSingleFrame function to AbstrRenderer.
I did debate appropriate places for this function. In the end,
AbstrRenderer (as a virtual method) was chosen. This function is
necessary to get the Lua batch renderer up and running.

Also added more verbose lua error output. It doesn't seem to function
as I would expect however. When the panic function is entered
it appears that all prior stack frames are lost, and we are not able to
retrieve any useful debug information. The code remains for future
reference.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2503 c36c8488-0289-0348-9b64-b301f74bd9a7