Create new performance query interface.
Get rid of a lot of the PH_ nonsense hacks.  There's now a perfcounter
interface which keeps everything.

If you want to add a new PerfCounter, you just need to add an enum
entry in Basics/PerfCounter.h.  Then increment it whenever appropriate
using the included 'StackTimer'.  The perf counter will be exported to
Lua, and resets itself every query.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2570 c36c8488-0289-0348-9b64-b301f74bd9a7