Split loading/registering data into two separate calls.
Mostly to change the registering to take a Dataset.  This means
that a derived renderer can play with the dataset that is
rendered, independently of controlling how the data are
loaded/created.

And that's now exactly what we do, now: both the 'rebricked' and
'normal' instances really just ask the IO manager to create the
Dataset object, and then use the register call (in this commit) to
register the dataset with the renderer.

Lua interface is unchanged.

git-svn-id: https://code.sci.utah.edu/svn/Tuvok@2582 c36c8488-0289-0348-9b64-b301f74bd9a7